remove unsafe usage of unsafe.Pointer()
Here we can define our callback with the GO pointer
GO will marshall the C pointer into the GO pointer when the callback
is called.
Using unsafe here makes the app crash with checks like `-race`
as the callback is not called from a GO thread, I guess the checkptr is
confused and does not find the associated go pointer.